Set OcpMaxVersion property in merge-index-image request in IIB
Refers to CLOUDDST-8769
