+++ iib/workers/tasks/utils.py
@@ -8,7 +8,9 @@ import json
 import logging
 import os
 import re
+import sqlite3
 import subprocess
+import time
 
 from retry import retry
 
@@ -19,12 +21,81 @@ from iib.workers.dogpile_cache import (
 )
 from operator_manifest.operator import ImageName
 
-from iib.exceptions import IIBError
+from iib.exceptions import IIBError, AddressAlreadyInUse
 from iib.workers.config import get_worker_config
 from iib.workers.api_utils import set_request_state
 
 log = logging.getLogger(__name__)
 dogpile_cache_region = create_dogpile_region()
+
+
+def _add_property_to_index(db_path, property):
+    """
+    Add a property to the index.
+
+    :param str db_path: path to the index database
+    :param dict property: a dict representing a property to be added to the index.db
+    """
+    insert = (
+        'INSERT INTO properties '
+        '(type, value, operatorbundle_name, operatorbundle_version, operatorbundle_path) '
+        'VALUES (?, ?, ?, ?, ?);'
+    )
+    con = sqlite3.connect(db_path)
+    # Insert property
+    con.execute(
+        insert,
+        (
+            property['type'],
+            property['value'],
+            property['operatorbundle_name'],
+            property['operatorbundle_version'],
+            property['operatorbundle_path'],
+        ),
+    )
+    con.commit()
+    con.close()
+
+
+def add_max_ocp_version_property(resolved_bundles, temp_dir):
+    """
+    Add the max ocp version property to bundles.
+
+    We need to ensure that any bundle which has deprecated/removed API(s) in 1.22/ocp 4.9
+    will have this property to prevent users from upgrading clusters to 4.9 before upgrading
+    the operator installed to a version that is compatible with 4.9
+
+    :param list resolved_bundles: list of resolved bundles to which the max ocp version property
+        will be added if missing
+    :param str temp_dir: directory location of the index image
+    """
+    # Get the CSV name and version (not just the bundle path)
+    db_path = temp_dir + "/database/index.db"
+    port, rpc_proc = serve_index_registry(db_path)
+
+    raw_bundles = run_cmd(
+        ['grpcurl', '-plaintext', f'localhost:{port}', 'api.Registry/ListBundles'],
+        exc_msg='Failed to get bundle data from index image',
+    )
+    rpc_proc.kill()
+
+    # Filter index image bundles to get pull spec for bundles in the request
+    updated_bundles = list(
+        filter(lambda b: b['bundlePath'] in resolved_bundles, get_bundle_json(raw_bundles))
+    )
+
+    for bundle in updated_bundles:
+        if _requires_max_ocp_version(bundle['bundlePath']):
+            log.info('adding property for %s', bundle['bundlePath'])
+            max_openshift_version_property = {
+                'type': 'olm.maxOpenShiftVersion',
+                'value': '4.8',
+                'operatorbundle_name': bundle['csvName'],
+                'operatorbundle_version': bundle['version'],
+                'operatorbundle_path': bundle['bundlePath'],
+            }
+            _add_property_to_index(db_path, max_openshift_version_property)
+            log.info('property added for %s', bundle['bundlePath'])
 
 
 def get_binary_image_from_config(ocp_version, distribution_scope, binary_image_config={}):
@@ -48,6 +119,136 @@ def get_binary_image_from_config(ocp_version, distribution_scope, binary_image_c
         )
 
     return binary_image
+
+
+def get_bundle_json(bundles):
+    """
+    Get bundle json from grpcurl response.
+
+    :param list bundles: response from grpcurl call to retrieve list of bundles in
+        an index
+    """
+    return [json.loads(bundle) for bundle in re.split(r'(?<=})\n(?={)', bundles)]
+
+
+def _requires_max_ocp_version(bundle):
+    """
+    Check if the bundle requires the olm.maxOpenShiftVersion property.
+
+    This property is required for bundles using deprecated APIs that don't already
+    have the olm.maxOpenShiftVersion property set.
+
+    :param str bundle: a string representing the bundle pull specification
+    :returns bool:
+    """
+    cmd = [
+        'operator-sdk',
+        'bundle',
+        'validate',
+        bundle,
+        '--select-optional',
+        'name=community',
+        '--output=json-alpha1',
+        '--image-builder',
+        'none',
+    ]
+    result = run_cmd(cmd, strict=False)
+    if result:
+        output = json.loads(result)
+        # check if the bundle validation failed
+        if not output['passed']:
+            # check if the failure is due to the presence of deprecated APIs
+            # and absence of the 'olm.maxOpenShiftVersion' property
+            # Note: there is no other error in the sdk that mentions this field
+            for msg in output['outputs']:
+                if 'olm.maxOpenShiftVersion' in msg['message']:
+                    return True
+    return False
+
+
+def serve_index_registry(db_path):
+    """
+    Locally start OPM registry service, which can be communicated with using gRPC queries.
+
+    Due to IIB's paralellism, the service can run multiple times, which could lead to port
+    binding conflicts. Resolution of port conflicts is handled in this function as well.
+
+    :param str db_path: path to index database containing the registry data.
+    :return: tuple containing port number of the running service and the running Popen object.
+    :rtype: (int, Popen)
+    :raises IIBError: if all tried ports are in use, or the command failed for another reason.
+    """
+    conf = get_worker_config()
+    port_start = conf['iib_grpc_start_port']
+    port_end = port_start + conf['iib_grpc_max_port_tries']
+
+    for port in range(port_start, port_end):
+        try:
+            return (
+                port,
+                serve_index_registry_at_port(
+                    db_path, port, conf['iib_grpc_max_tries'], conf['iib_grpc_init_wait_time']
+                ),
+            )
+        except AddressAlreadyInUse:
+            log.info('Port %d is in use, trying another.', port)
+
+    err_msg = f'No free port has been found after {conf.get("iib_grpc_max_port_tries")} attempts.'
+    log.error(err_msg)
+    raise IIBError(err_msg)
+
+
+@retry(exceptions=IIBError, tries=2, logger=log)
+def serve_index_registry_at_port(db_path, port, max_tries, wait_time):
+    """
+    Start an image registry service at a specified port.
+
+    :param str db_path: path to index database containing the registry data.
+    :param str int port: port to start the service on.
+    :param max_tries: how many times to try to start the service before giving up.
+    :param wait_time: time to wait before checking if the service is initialized.
+    :return: object of the running Popen process.
+    :rtype: Popen
+    :raises IIBError: if the process has failed to initialize too many times, or an unexpected
+        error occured.
+    :raises AddressAlreadyInUse: if the specified port is already being used by another service.
+    """
+    cmd = ['opm', 'registry', 'serve', '-p', str(port), '-d', db_path, '-t', '/dev/null']
+    for attempt in range(max_tries):
+        rpc_proc = subprocess.Popen(
+            cmd,
+            cwd=os.path.dirname(db_path),
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            universal_newlines=True,
+        )
+        start_time = time.time()
+        while time.time() - start_time < wait_time:
+            time.sleep(1)
+            ret = rpc_proc.poll()
+            # process has terminated
+            if ret is not None:
+                stderr = rpc_proc.stderr.read()
+                if 'address already in use' in stderr:
+                    raise AddressAlreadyInUse(f'Port {port} is already used by a different service')
+                raise IIBError(f'Command "{" ".join(cmd)}" has failed with error "{stderr}"')
+
+            # query the service to see if it has started
+            try:
+                output = run_cmd(
+                    ['grpcurl', '-plaintext', f'localhost:{port}', 'list', 'api.Registry']
+                )
+            except IIBError:
+                output = ''
+
+            if 'api.Registry.ListBundles' in output or 'api.Registry.ListPackages' in output:
+                log.debug('Started the command "%s"', ' '.join(cmd))
+                log.info('Index registry service has been initialized.')
+                return rpc_proc
+
+        rpc_proc.kill()
+
+    raise IIBError(f'Index registry has not been initialized after {max_tries} tries')
 
 
 class RequestConfig:
@@ -875,11 +1076,11 @@ def grpcurl_get_db_data(from_index, base_dir, endpoint):
     # This is temporary solution till we refactor the code and move those functions to utils
     from iib.workers.tasks.build import (
         _get_index_database,
-        _serve_index_registry,
+        serve_index_registry,
     )
 
     db_path = _get_index_database(from_index, base_dir)
-    port, rpc_proc = _serve_index_registry(db_path)
+    port, rpc_proc = serve_index_registry(db_path)
 
     if endpoint not in ["api.Registry/ListPackages", "api.Registry/ListBundles"]:
         raise IIBError(f"The endpoint '{endpoint}' is not allowed to be used")
