+++ tests/test_workers/test_tasks/test_utils.py
@@ -981,7 +981,7 @@ def test_get_binary_image_config_no_config_val():
     'endpoint', ("api.Registry/ListPackages", "api.Registry/ListBundles",),
 )
 @mock.patch('iib.workers.tasks.utils.run_cmd')
-@mock.patch('iib.workers.tasks.build._serve_index_registry')
+@mock.patch('iib.workers.tasks.build.serve_index_registry')
 @mock.patch('iib.workers.tasks.build._get_index_database')
 def test_grpcurl_get_db_data_success(mock_gid, mock_sir, mock_run_cmd, tmpdir, endpoint):
 
@@ -1003,7 +1003,7 @@ def test_grpcurl_get_db_data_success(mock_gid, mock_sir, mock_run_cmd, tmpdir, e
     ),
 )
 @mock.patch('iib.workers.tasks.utils.run_cmd')
-@mock.patch('iib.workers.tasks.build._serve_index_registry')
+@mock.patch('iib.workers.tasks.build.serve_index_registry')
 @mock.patch('iib.workers.tasks.build._get_index_database')
 def test_grpcurl_get_db_data_wrong_endpoint(
     mock_gid, mock_sir, mock_run_cmd, tmpdir, endpoint, err_msg
@@ -1019,3 +1019,40 @@ def test_grpcurl_get_db_data_wrong_endpoint(
     mock_sir.assert_called_once()
     mock_gid.assert_called_once()
     mock_run_cmd.assert_not_called()
+
+
+@mock.patch('time.sleep')
+@mock.patch('subprocess.Popen')
+@mock.patch('iib.workers.tasks.utils.run_cmd')
+def test_serve_image_registry(mock_run_cmd, mock_popen, mock_sleep, tmpdir):
+    my_mock = mock.MagicMock()
+    mock_popen.return_value = my_mock
+    my_mock.stderr.read.side_effect = ['address already in use', 'address already in use']
+    mock_run_cmd.return_value = 'api.Registry.ListBundles'
+    my_mock.poll.side_effect = [1, 1, None]
+    port, _ = utils.serve_index_registry('some_path.db')
+    assert port == 50053
+    assert my_mock.poll.call_count == 3
+
+
+@mock.patch('iib.workers.tasks.utils.get_worker_config')
+@mock.patch('time.sleep')
+@mock.patch('subprocess.Popen')
+def test_serve_image_registry_no_ports(mock_popen, mock_sleep, mock_config, tmpdir):
+    my_mock = mock.MagicMock()
+    mock_popen.return_value = my_mock
+    my_mock.stderr.read.side_effect = [
+        'address already in use',
+        'address already in use',
+        'address already in use',
+    ]
+    my_mock.poll.side_effect = [1, 1, 1, None]
+    mock_config.return_value = {
+        'iib_grpc_start_port': 50051,
+        'iib_grpc_init_wait_time': 1,
+        'iib_grpc_max_port_tries': 3,
+        'iib_grpc_max_tries': 3,
+    }
+    with pytest.raises(IIBError, match='No free port has been found after 3 attempts.'):
+        utils.serve_index_registry('some_path.db')
+    assert my_mock.poll.call_count == 3
